add scenario action-deploy, delete, update, check
